Project structure refactor
