+++ app/src/main/java/edu/agh/susgame/front/service/web/WebGameService.kt
@@ -4,7 +4,7 @@ import edu.agh.susgame.dto.rest.model.LobbyId
 import edu.agh.susgame.dto.rest.model.PlayerNickname
 import edu.agh.susgame.dto.socket.ClientSocketMessage
 import edu.agh.susgame.front.service.interfaces.GameService
-import edu.agh.susgame.front.service.web.rest.AbstractRest
+import edu.agh.susgame.front.rest.AbstractRest
 import edu.agh.susgame.front.service.web.socket.GameWebSocketListener
 import edu.agh.susgame.front.util.AppConfig
 import kotlinx.coroutines.CoroutineScope
